Don't mark a patch transform as set if the parent transform is not set.
When updating a patch from another artist, if the other artist's
transform is marked as "uninitialized" (`_transformSet == False`), then
don't mark the updatee's transform as set either.

Typically, this occurs if the updator has not been added to an Axes yet;
its transform will be set when that adding occurs.  In that case, the
updatee's transform also needs to be set when actually being added to an
Axes.