Only set the wait cursor if the last draw was >1s ago.
This avoids setting the wait cursor when panning or zooming (that
behavior was somewhat annoying), or during an animation that renders at
least one frame per second.

Rename `_set_cursor` to `_update_cursor` as it is fairly different from
`set_cursor` and the previous name could cause confusion.  Move the
defintion of cursor-related methods to a single place.

In the case of gtk, just strip out the offending cursor related code as
the backend is deprecated and slated for removal anyways.