Remove "experimental" fontconfig font_manager backend.
The "experimental" fontconfig backend for font_manager was never
publicly accessible: even if one does
`matplotlib.font_manager.USE_FONTCONFIG = True`, that didn't change the
already existing `findfont`.  The only way to access it was to edit the
source, or use `reload()`, neither of which really count as public
API...

Note that our "use" of "fontconfig-like" patterns actually has subtly
different semantics from actual fontconfig patterns, so it's not as if
that backend was correctly working anyways.

We don't need to set `fontManager.default_size` when loading it as
we already check that fontManager has the correct `__version__`
(and thus must have a correct `default_size`; moreover the only
use of `default_size` is in `sizeval1 = self.default_size *
font_scalings[size1]` so it's not as if a value of `None` (if it had
somehow been missing) was going to be helpful anyways...

`get_cachedir()` always returns a real directory (creating a temporary
one if necessary), so we can drop the code paths handling
`get_cachedir() == None`.

Note that we still rely on fontconfig to *list* fonts; the backend only
added an (non-publicly-accessible, per above) option to *match* fonts
using fontconfig.