Simplify cleanup decorator implementation.
Introduce a single private `_cleanup_cm` contextmanager and use it to
implement `CleanupTestCase` and `@cleanup`.

Use `warnings.catch_warnings` to avoid completely destroying a
preexisting warnings filter, instead just restoring the filter that
existed before the test started.

Use `matplotlib.style.context` to restore the style at exit, as it
relies on rc_context which is ultimately more efficient than
`rcParams.update` as it skips revalidation.

Deprecate CleanupTest (and implement it in terms of CleanupTestCase), as
it is clearly a nose-oriented base class that could have been deprecated
at the same time as ImageComparisonTest.